cookbook/financial_planner/requirements.txt
pydantic-ai==0.0.14
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.18
requests>=2.31.0
python-dateutil>=2.8.2
scipy>=1.11.0